Create AnthonyFloresProjectOne.pdf
For the first project, I created three organic sculptural renderings in Maya. My original concept was to create a ghost-like piece that would be entangled within a wall space. With the use of Arnold lighting in addition to a Phong material, I gave my piece a glowing feature. In terms of modularity, the three pieces work together to create one image, but they can be seen independently if viewed.  Since I was not able to create a transparent fog effect, I added three different ghost image textures to all the sculptures to give more of a ghost effect.

Google Docs link
https://drive.google.com/file/d/0B2QJ5u2ea9V5Sjd6aTY2MXZza0k/view?usp=sharing